+++ src3/arvoreB.h
@@ -12,11 +12,11 @@ class ArvoreB{
         // Função para percorrer a árvore B
         void percorrerArvore();
         // Função para procurar a chave k na árvore B
-        NoB* buscarChave(int k, unsigned *numTrocas, unsigned *numComparacoes);
+        NoB* buscarChave(int k, unsigned long long *numTrocas, unsigned long long *numComparacoes);
         // Função para inserir uma chave na árvore B
-        void inserirChave(int k, unsigned *numTrocas, unsigned *numComparacoes);
+        void inserirChave(int k, unsigned long long *numTrocas, unsigned long long *numComparacoes);
         // Remove a chave k da árvore B 
-        void removerChave(int k, unsigned *numTrocas, unsigned *numComparacoes); 
+        void removerChave(int k, unsigned long long *numTrocas, unsigned long long *numComparacoes); 
 
 };
